avia_ticket_chatbot/requirements.txt
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
coverage==5.0.3
decorator==4.4.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.9
pydevd-pycharm==192.6262.63
pygame==1.9.6
Pygments==2.4.2
requests==2.22.0
simple-draw==2.6.5
six==1.12.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.25.7
vk-api==11.6.1
wcwidth==0.1.7
